Create a function to return the computer choice
I created a function `getComputerChoice` that randomly selects
between 'rock', 'paper', and 'scissors'. This function will be used
in the game logic to determine the computer's choice during each round
of the game.

To help with the game logic, I also added a simple random selection
mechanism using `Math.random()` to ensure that the computer's choice
is unpredictable.