Create function to get the player's choice
I created a function that prompts the user to enter their choice of
rock, paper, or scissors. This function is used to get the player's
choice before the game starts. It does not currently do any validation
on the input, so it will accept any string.

I also added a variable to store the player's choice as a lowercase
string, which will be used later in the game logic.