Update number of rounds
I changed the logic of the game so that iot only plays five rounds
total. After the fifth round, the final winner is shown and the
total scores and round numbers are reset to 0.  If the
user clicks one of the buttons again, the game remove the final winner
text and restarts the game on round 1.